+++ solvers/BUILD.bazel
@@ -873,8 +873,30 @@ drake_cc_variant_library(
         ":mathematical_program",
     ],
     deps_enabled = [
+        ":mosek_solver_internal",
         "//common:scoped_singleton",
         "//common:scope_exit",
+        "@mosek",
+    ],
+)
+
+# If --config mosek is used, this target builds object code constructs internal
+# functions used in mosek_solver. Otherwise, this target builds no object
+# code.
+drake_cc_variant_library(
+    name = "mosek_solver_internal",
+    opt_in_condition = "//tools:with_mosek",
+    srcs_always = [],
+    srcs_enabled = ["mosek_solver_internal.cc"],
+    srcs_disabled = [],
+    hdrs = ["mosek_solver_internal.h"],
+    internal = True,
+    visibility = ["//visibility:private"],
+    interface_deps = [":mathematical_program_result"],
+    deps_always = [],
+    deps_enabled = [
+        ":aggregate_costs_constraints",
+        ":mathematical_program",
         "//math:quadratic_form",
         "@mosek",
     ],
